move report api dynamic tests to unittest
