+++ test/unittest/routes/reporter_api/reporter_api_test.go
@@ -5,11 +5,14 @@ import (
 	"fmt"
 	"net/http"
 	"net/http/httptest"
+	"soarca/internal/reporter/downstream_reporter/cache"
 	api_model "soarca/models/api"
+	"soarca/models/cacao"
 	cache_model "soarca/models/cache"
 	"soarca/routes/reporter"
-	mock_cache "soarca/test/unittest/mocks/mock_cache"
+	mock_time "soarca/test/unittest/mocks/mock_utils/time"
 	"testing"
+	"time"
 
 	"github.com/google/uuid"
 
@@ -17,15 +20,119 @@ import (
 	"github.com/go-playground/assert/v2"
 )
 
-func TestGetExecutionsInvocation(t *testing.T) {
-	mock_cache_reporter := &mock_cache.Mock_Cache{}
-	mock_cache_reporter.On("GetExecutions").Return([]cache_model.ExecutionEntry{}, nil)
+func TestGetExecutions(t *testing.T) {
+
+	mock_time := new(mock_time.MockTime)
+	cacheReporter := cache.New(mock_time, 10)
+
+	expectedCommand := cacao.Command{
+		Type:    "ssh",
+		Command: "ssh ls -la",
+	}
+
+	expectedVariables := cacao.Variable{
+		Type:  "string",
+		Name:  "var1",
+		Value: "testing",
+	}
+
+	step1 := cacao.Step{
+		Type:          "action",
+		ID:            "action--test",
+		Name:          "ssh-tests",
+		StepVariables: cacao.NewVariables(expectedVariables),
+		Commands:      []cacao.Command{expectedCommand},
+		Cases:         map[string]string{},
+		OnCompletion:  "end--test",
+		Agent:         "agent1",
+		Targets:       []string{"target1"},
+	}
+
+	end := cacao.Step{
+		Type: "end",
+		ID:   "end--test",
+		Name: "end step",
+	}
+
+	expectedAuth := cacao.AuthenticationInformation{
+		Name: "user",
+		ID:   "auth1",
+	}
+
+	expectedTarget := cacao.AgentTarget{
+		Name:               "sometarget",
+		AuthInfoIdentifier: "auth1",
+		ID:                 "target1",
+	}
+
+	expectedAgent := cacao.AgentTarget{
+		Type: "soarca",
+		Name: "soarca-ssh",
+	}
+
+	playbook := cacao.Playbook{
+		ID:                            "test",
+		Type:                          "test",
+		Name:                          "ssh-test",
+		WorkflowStart:                 step1.ID,
+		AuthenticationInfoDefinitions: map[string]cacao.AuthenticationInformation{"id": expectedAuth},
+		AgentDefinitions:              map[string]cacao.AgentTarget{"agent1": expectedAgent},
+		TargetDefinitions:             map[string]cacao.AgentTarget{"target1": expectedTarget},
+
+		Workflow: map[string]cacao.Step{step1.ID: step1, end.ID: end},
+	}
+	executionId0 := uuid.MustParse("6ba7b810-9dad-11d1-80b4-00c04fd430c0")
+	executionId1 := uuid.MustParse("6ba7b810-9dad-11d1-80b4-00c04fd430c1")
+	executionId2 := uuid.MustParse("6ba7b810-9dad-11d1-80b4-00c04fd430c2")
+
+	executionIds := []uuid.UUID{
+		executionId0,
+		executionId1,
+		executionId2,
+	}
+
+	layout := "2006-01-02T15:04:05.000Z"
+	str := "2014-11-12T11:45:26.371Z"
+	timeNow, _ := time.Parse(layout, str)
+	mock_time.On("Now").Return(timeNow)
+
+	expectedStarted, _ := time.Parse(layout, str)
+	expectedEnded, _ := time.Parse(layout, "0001-01-01T00:00:00Z")
+
+	expectedExecutions := []cache_model.ExecutionEntry{}
+	for _, executionId := range executionIds {
+		t.Log(executionId)
+		entry := cache_model.ExecutionEntry{
+			ExecutionId:    executionId,
+			PlaybookId:     "test",
+			Started:        expectedStarted,
+			Ended:          expectedEnded,
+			StepResults:    map[string]cache_model.StepResult{},
+			PlaybookResult: nil,
+			Status:         2,
+		}
+		expectedExecutions = append(expectedExecutions, entry)
+	}
+
+	err := cacheReporter.ReportWorkflowStart(executionId0, playbook)
+	if err != nil {
+		t.Fail()
+	}
+
+	err = cacheReporter.ReportWorkflowStart(executionId1, playbook)
+	if err != nil {
+		t.Fail()
+	}
+	err = cacheReporter.ReportWorkflowStart(executionId2, playbook)
+	if err != nil {
+		t.Fail()
+	}
 
 	app := gin.New()
 	gin.SetMode(gin.DebugMode)
 
 	recorder := httptest.NewRecorder()
-	reporter.Routes(app, mock_cache_reporter)
+	reporter.Routes(app, cacheReporter)
 
 	request, err := http.NewRequest("GET", "/reporter/", nil)
 	if err != nil {
@@ -33,66 +140,119 @@ func TestGetExecutionsInvocation(t *testing.T) {
 	}
 
 	app.ServeHTTP(recorder, request)
-	expectedString := "[]"
+	expectedByte, err := json.Marshal(expectedExecutions)
+	if err != nil {
+		t.Log("failed to decode expected struct to json")
+		t.Fail()
+	}
+	expectedString := string(expectedByte)
 	assert.Equal(t, expectedString, recorder.Body.String())
 	assert.Equal(t, 200, recorder.Code)
 
-	mock_cache_reporter.AssertExpectations(t)
+	mock_time.AssertExpectations(t)
 }
 
-func TestGetExecutionReportInvocation(t *testing.T) {
-	mock_cache_reporter := &mock_cache.Mock_Cache{}
+func TestGetExecutionReport(t *testing.T) {
+	// Create real cache, create real reporter api object
+	// Do executions, test retrieval via api
+
+	mock_time := new(mock_time.MockTime)
+	cacheReporter := cache.New(mock_time, 10)
+
+	expectedCommand := cacao.Command{
+		Type:    "ssh",
+		Command: "ssh ls -la",
+	}
+
+	expectedVariables := cacao.Variable{
+		Type:  "string",
+		Name:  "var1",
+		Value: "testing",
+	}
+
+	step1 := cacao.Step{
+		Type:          "action",
+		ID:            "action--test",
+		Name:          "ssh-tests",
+		StepVariables: cacao.NewVariables(expectedVariables),
+		Commands:      []cacao.Command{expectedCommand},
+		Cases:         map[string]string{},
+		OnCompletion:  "end--test",
+		Agent:         "agent1",
+		Targets:       []string{"target1"},
+	}
+
+	end := cacao.Step{
+		Type: "end",
+		ID:   "end--test",
+		Name: "end step",
+	}
+
+	expectedAuth := cacao.AuthenticationInformation{
+		Name: "user",
+		ID:   "auth1",
+	}
+
+	expectedTarget := cacao.AgentTarget{
+		Name:               "sometarget",
+		AuthInfoIdentifier: "auth1",
+		ID:                 "target1",
+	}
+
+	expectedAgent := cacao.AgentTarget{
+		Type: "soarca",
+		Name: "soarca-ssh",
+	}
+
+	playbook := cacao.Playbook{
+		ID:                            "test",
+		Type:                          "test",
+		Name:                          "ssh-test",
+		WorkflowStart:                 step1.ID,
+		AuthenticationInfoDefinitions: map[string]cacao.AuthenticationInformation{"id": expectedAuth},
+		AgentDefinitions:              map[string]cacao.AgentTarget{"agent1": expectedAgent},
+		TargetDefinitions:             map[string]cacao.AgentTarget{"target1": expectedTarget},
+
+		Workflow: map[string]cacao.Step{step1.ID: step1, end.ID: end},
+	}
+	executionId0 := uuid.MustParse("6ba7b810-9dad-11d1-80b4-00c04fd430c0")
+	executionId1 := uuid.MustParse("6ba7b810-9dad-11d1-80b4-00c04fd430c1")
+	executionId2 := uuid.MustParse("6ba7b810-9dad-11d1-80b4-00c04fd430c2")
+
+	layout := "2006-01-02T15:04:05.000Z"
+	str := "2014-11-12T11:45:26.371Z"
+	timeNow, _ := time.Parse(layout, str)
+	mock_time.On("Now").Return(timeNow)
+
+	err := cacheReporter.ReportWorkflowStart(executionId0, playbook)
+	if err != nil {
+		t.Fail()
+	}
+	err = cacheReporter.ReportStepStart(executionId0, step1, cacao.NewVariables(expectedVariables))
+	if err != nil {
+		t.Fail()
+	}
+
+	err = cacheReporter.ReportWorkflowStart(executionId1, playbook)
+	if err != nil {
+		t.Fail()
+	}
+	err = cacheReporter.ReportWorkflowStart(executionId2, playbook)
+	if err != nil {
+		t.Fail()
+	}
+	err = cacheReporter.ReportStepEnd(executionId0, step1, cacao.NewVariables(expectedVariables), nil)
+	if err != nil {
+		t.Fail()
+	}
+
 	app := gin.New()
 	gin.SetMode(gin.DebugMode)
 
 	recorder := httptest.NewRecorder()
-	reporter.Routes(app, mock_cache_reporter)
+	reporter.Routes(app, cacheReporter)
 
-	executionId0, _ := uuid.Parse("6ba7b810-9dad-11d1-80b4-00c04fd430c0")
-
-	expectedCache := `{
-		"ExecutionId":"6ba7b810-9dad-11d1-80b4-00c04fd430c0",
-		"PlaybookId":"test",
-		"Started":"2014-11-12T11:45:26.371Z",
-		"Ended":"0001-01-01T00:00:00Z",
-		"StepResults":{
-		   "action--test":{
-			  "ExecutionId":"6ba7b810-9dad-11d1-80b4-00c04fd430c0",
-			  "StepId":"action--test",
-			  "Started":"2014-11-12T11:45:26.371Z",
-			  "Ended":"2014-11-12T11:45:26.371Z",
-			  "Variables":{
-				 "var1":{
-					"type":"string",
-					"name":"var1",
-					"value":"testing"
-				 }
-			  },
-			  "Status":0,
-			  "Error":null
-		   }
-		},
-		"PlaybookResult":null,
-		"Status":2
-	 }`
-	expectedCacheData := cache_model.ExecutionEntry{}
-	err := json.Unmarshal([]byte(expectedCache), &expectedCacheData)
-	if err != nil {
-		t.Log(err)
-		t.Log("Could not parse data to JSON")
-		t.Fail()
-	}
-
-	mock_cache_reporter.On("GetExecutionReport", executionId0).Return(expectedCacheData, nil)
-
-	request, err := http.NewRequest("GET", fmt.Sprintf("/reporter/%s", executionId0), nil)
-	if err != nil {
-		t.Log(err)
-		t.Fail()
-	}
-	app.ServeHTTP(recorder, request)
-
-	expectedResponse := `{
+	expected := `{
 		"Type":"execution_status",
 		"ExecutionId":"6ba7b810-9dad-11d1-80b4-00c04fd430c0",
 		"PlaybookId":"test",
@@ -121,13 +281,20 @@ func TestGetExecutionReportInvocation(t *testing.T) {
 		"Error":"",
 		"RequestInterval":5
 	}`
-	expectedResponseData := api_model.PlaybookExecutionReport{}
-	err = json.Unmarshal([]byte(expectedResponse), &expectedResponseData)
+	expectedData := api_model.PlaybookExecutionReport{}
+	err = json.Unmarshal([]byte(expected), &expectedData)
 	if err != nil {
 		t.Log(err)
 		t.Log("Could not parse data to JSON")
 		t.Fail()
 	}
+
+	request, err := http.NewRequest("GET", fmt.Sprintf("/reporter/%s", executionId0), nil)
+	if err != nil {
+		t.Log(err)
+		t.Fail()
+	}
+	app.ServeHTTP(recorder, request)
 
 	receivedData := api_model.PlaybookExecutionReport{}
 	err = json.Unmarshal(recorder.Body.Bytes(), &receivedData)
@@ -137,6 +304,7 @@ func TestGetExecutionReportInvocation(t *testing.T) {
 		t.Fail()
 	}
 
-	assert.Equal(t, expectedResponseData, receivedData)
-	mock_cache_reporter.AssertExpectations(t)
+	assert.Equal(t, expectedData, receivedData)
+
+	mock_time.AssertExpectations(t)
 }
